Split sys into platform specific files
